Fix bug in reading data when a dimension has only one value
When a dimension has only one value it needs to be wrapped in an array so
the length used when computing the offset is the one of the array and not
the value string itself.

Adds splitting the data also by “;”, “,” and tabs (according to the pc-axis spec),
instead of only by spaces.